Refactoring budget table to its own file
